refactor(): incremental changes

+++ gateways/core/gcp-pubsub/start.go
@@ -29,8 +29,9 @@ import (
 	"google.golang.org/api/option"
 )
 
-// EventSourceListener implements Eventing
+// EventSourceListener implements Eventing for gcp pub-sub event source
 type EventSourceListener struct {
+	// Logger to log stuff
 	Logger *logrus.Logger
 }
 
@@ -52,23 +53,36 @@ func (listener *EventSourceListener) StartEventSource(eventSource *gateways.Even
 
 // listenEvents listens to GCP PubSub events
 func (listener *EventSourceListener) listenEvents(ctx context.Context, eventSource *gateways.EventSource, dataCh chan []byte, errorCh chan error, doneCh chan struct{}) {
+	// In order to listen events from GCP PubSub,
+	// 1. Parse the event source that contains configuration to connect to GCP PubSub
+	// 2. Create a new PubSub client
+	// 3. Create the topic if one doesn't exist already
+	// 4. Create a subscription if one doesn't exist already.
+	// 5. Start listening to messages on the queue
+	// 6. Once the event source is stopped perform cleaning up - 1. Delete the subscription if configured so 2. Close the PubSub client
+
+	logger := listener.Logger.WithField(common.LabelEventSource, eventSource.Name)
+
+	logger.Infoln("parsing PubSub event source...")
 	var pubsubEventSource *v1alpha1.PubSubEventSource
 	if err := yaml.Unmarshal(eventSource.Value, &pubsubEventSource); err != nil {
-		listener.Logger.WithError(err).WithField(common.LabelEventSource, eventSource.Name).Errorln("failed to parse the event source")
 		errorCh <- err
 		return
 	}
 
-	// Create a new topic with the given name if none exists
-	logger := listener.Logger.WithField(common.LabelEventSource, eventSource.Name).WithField("topic", pubsubEventSource.Topic)
+	logger = logger.WithField("topic", pubsubEventSource.Topic)
 
+	logger.Infoln("setting up a client to connect to PubSub...")
+
+	// Create a new topic with the given name if none exists
 	client, err := pubsub.NewClient(ctx, pubsubEventSource.ProjectID, option.WithCredentialsFile(pubsubEventSource.CredentialsFile))
 	if err != nil {
 		errorCh <- err
 		return
 	}
 
-	topicClient := client // use same client for topic and subscription by default
+	// use same client for topic and subscription by default
+	topicClient := client
 	if pubsubEventSource.TopicProjectID != "" && pubsubEventSource.TopicProjectID != pubsubEventSource.ProjectID {
 		topicClient, err = pubsub.NewClient(ctx, pubsubEventSource.TopicProjectID, option.WithCredentialsFile(pubsubEventSource.CredentialsFile))
 		if err != nil {
@@ -77,6 +91,7 @@ func (listener *EventSourceListener) listenEvents(ctx context.Context, eventSour
 		}
 	}
 
+	logger.Infoln("getting topic information from PubSub...")
 	topic := topicClient.Topic(pubsubEventSource.Topic)
 	exists, err := topic.Exists(ctx)
 	if err != nil {
@@ -84,15 +99,18 @@ func (listener *EventSourceListener) listenEvents(ctx context.Context, eventSour
 		return
 	}
 	if !exists {
-		logger.Info("Creating GCP PubSub topic")
+		logger.Infoln("topic doesn't exist, creating the GCP PubSub topic...")
 		if _, err := topicClient.CreateTopic(ctx, pubsubEventSource.Topic); err != nil {
 			errorCh <- err
 			return
 		}
 	}
 
-	logger.Info("Subscribing to GCP PubSub topic")
 	subscriptionName := fmt.Sprintf("%s-%s", eventSource.Name, eventSource.Id)
+
+	logger = logger.WithField("subscription", subscriptionName)
+
+	logger.Infoln("subscribing to PubSub topic...")
 	subscription := client.Subscription(subscriptionName)
 	exists, err = subscription.Exists(ctx)
 
@@ -101,15 +119,16 @@ func (listener *EventSourceListener) listenEvents(ctx context.Context, eventSour
 		return
 	}
 	if exists {
-		logger.Warn("Using an existing subscription")
+		logger.Warnln("using an existing subscription...")
 	} else {
-		logger.Info("Creating subscription")
+		logger.Infoln("creating a new subscription...")
 		if _, err := client.CreateSubscription(ctx, subscriptionName, pubsub.SubscriptionConfig{Topic: topic}); err != nil {
 			errorCh <- err
 			return
 		}
 	}
 
+	logger.Infoln("listening for messages from PubSub...")
 	err = subscription.Receive(ctx, func(msgCtx context.Context, m *pubsub.Message) {
 		logger.Info("received GCP PubSub Message from topic")
 		dataCh <- m.Data
@@ -122,14 +141,15 @@ func (listener *EventSourceListener) listenEvents(ctx context.Context, eventSour
 
 	<-doneCh
 
-	// after this point, panic on errors
-	logger.Info("deleting GCP PubSub subscription")
-	if err = subscription.Delete(context.Background()); err != nil {
-		panic(err)
+	if pubsubEventSource.DeleteSubscriptionOnFinish {
+		logger.Info("deleting PubSub subscription...")
+		if err = subscription.Delete(context.Background()); err != nil {
+			logger.WithError(err).Errorln("failed to delete the PubSub subscription")
+		}
 	}
 
-	logger.Info("closing GCP PubSub client")
+	logger.Info("closing PubSub client...")
 	if err = client.Close(); err != nil {
-		panic(err)
+		logger.WithError(err).Errorln("failed to close the PubSub client")
 	}
 }
